x86: reduce irq_cpustat_t's __softirq_pending to 32 bits
Assembly code was already only accessing the low 32 bits of it, and
we're far away from using all 32 bits of it.

Noticed-by: Andrew Cooper <andrew.cooper3@citrix.com>
Acked-by: Keir Fraser <keir@xen.org>

+++ xen/include/asm-x86/hardirq.h
@@ -5,7 +5,7 @@
 #include <xen/cache.h>
 
 typedef struct {
-	unsigned long __softirq_pending;
+	unsigned int __softirq_pending;
 	unsigned int __local_irq_count;
 	unsigned int __nmi_count;
 	bool_t __mwait_wakeup;
